requirements.txt
rich>=13.0
requests>=2.31
Pillow>=10.0
